update vacany classess

+++ src/main/java/com/example/community/Controller/VacancyController.java
@@ -1,4 +1,5 @@
 package com.example.community.Controller;
+
 import com.example.community.model.Vacancy;
 import com.example.community.Exception.ResourceNotFoundException;
 import com.example.community.repository.VacancyRepository;
@@ -7,29 +8,32 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import java.util.List;
+
 @RestController
 @RequestMapping("/posts/vacancy")
 public class VacancyController {
 
-@Autowired
+    @Autowired
     VacancyRepository vacancyRepo;
 
-@GetMapping("/all")
-public List<Vacancy> getAllPosts() {
+    @GetMapping("/all")
+    public List<Vacancy> getAllPosts() {
         return vacancyRepo.findAll();
-        }
-@PostMapping("/add")
-public Vacancy createNote(@Valid @RequestBody Vacancy post) {
+    }
+
+    @PostMapping("/add")
+    public Vacancy createNote(@Valid @RequestBody Vacancy post) {
         return vacancyRepo.save(post);
-        }
-// Get a Single Note
-@GetMapping("/search/{userID,postID}")
-public Vacancy getVacancyById(@PathVariable(value = "postID") Long postId,@PathVariable(value = "userID") Long userId) {
+    }
+
+    @GetMapping("/search/{userID,postID}")
+    public Vacancy getVacancyById(@PathVariable(value = "postID") Long postId, @PathVariable(value = "userID") Long userId) {
         return vacancyRepo.findById(postId)
-        .orElseThrow(() -> new ResourceNotFoundException("Post", "id", postId));
-        }
-@PutMapping("/{postid}")
-public Vacancy updateNote(@PathVariable(value = "postid") Long postId, @Valid @RequestBody Vacancy postDetails) {
+                .orElseThrow(() -> new ResourceNotFoundException("Post", "id", postId));
+    }
+
+    @PutMapping("/{postid}")
+    public Vacancy updatevacancy(@PathVariable(value = "postid") Long postId, @Valid @RequestBody Vacancy postDetails) {
         Vacancy post = vacancyRepo.findById(postId).orElseThrow(() -> new ResourceNotFoundException("Post", "id", postId));
         post.setDescreption(postDetails.getDescreption());
         post.setAppid(postDetails.getAppid());
@@ -40,15 +44,26 @@ public Vacancy updateNote(@PathVariable(value = "postid") Long postId, @Valid @R
         post.setTilte(postDetails.getTitle());
         post.setSalary(postDetails.getSalary());
         post.setType(postDetails.getType());
-        Vacancy updatedNote = vacancyRepo.save(post);
-        return updatedNote;
-        }
-@DeleteMapping("/{postid,userid}")
-public ResponseEntity<?> deletepost(@PathVariable(value = "postid") Long postId,@PathVariable(value = "userid") Long userId) {
+        Vacancy updatedvacancy = vacancyRepo.save(post);
+        return updatedvacancy;
+    }
+
+    @DeleteMapping("/{postid,companyid}")
+    public ResponseEntity<?> deletevacancy(@PathVariable(value = "postid") Long postId, @PathVariable(value = "companyid") String companyId) {
         Vacancy post = vacancyRepo.findById(postId)
-        .orElseThrow(() -> new ResourceNotFoundException("Post", "id", postId));
-        vacancyRepo.delete(post);
-        return ResponseEntity.ok().build();
+                .orElseThrow(() -> new ResourceNotFoundException("Post", "id", postId));
+        if (post.getPosterID().equals(companyId)) {
+            vacancyRepo.delete(post);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.notFound().build();
+    }
+    @PutMapping("/{postid,vote}")
+        public Vacancy VoteToVacancy(@PathVariable(value = "postid") Long postId,@PathVariable(value = "vote") int vote, @Valid @RequestBody Vacancy postDetails) {
+            Vacancy post = vacancyRepo.findById(postId).orElseThrow(() -> new ResourceNotFoundException("Post", "id", postId));
+            post.setvote(vote);
+            Vacancy updatedVacancy = vacancyRepo.save(post);
+            return updatedVacancy;
         }
 
-        }
+}
